FY-166/refactor: rename GameSession -> SubGameSession & pack into module
