+++ src/livegame/SubGameSession/BallTrack.py
@@ -3,7 +3,7 @@ from typing import List, Tuple
 from enum import Enum
 
 from livegame.GameConfig import GameConfig
-from livegame.BallTrackSegment import (
+from livegame.SubGameSession.BallTrackSegment import (
     BallTrackSegment,
     get_ball_track_segment_to_wall,
     get_ball_track_segment_to_paddle,
